Use long for {Primitive}BigArray#capacity
Previous implementations supported accesses by long index, but used
int as the capacity field which could overflow. Overflowing was mostly
benign but would cause #ensureCapacity(long capacity) to proceed down
the slow-path unnecessarily.

+++ presto-array/src/main/java/com/facebook/presto/array/BooleanBigArray.java
@@ -34,7 +34,7 @@ public final class BooleanBigArray
     private final boolean initialValue;
 
     private boolean[][] array;
-    private int capacity;
+    private long capacity;
     private int segments;
 
     /**
